refactor: Comment Repository를 interface로 분리
